Double Precison working version

+++ src/visualisation.h
@@ -4,6 +4,6 @@ extern "C" void display();
 
 extern "C" bool initialize(int *argc, char **argv, int v, int f);
 
-extern "C" void set_fn(float2 *dev_fn);
+extern "C" void set_fn(double2 *dev_fn);
 
 extern "C" void initiateVBOData(SimpleTriMesh mesh);
